Revert "Try and implement a ViewPager"
This reverts commit fda8d0c8b35cbcb794831dfa313573cba090a8f7.
