--- app/src/main/res/layout/frag_absence_list.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <LinearLayout
-        android:id="@+id/noAbsenceContainer"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_below="@+id/imageButton"
-        android:gravity="center_vertical|center_horizontal"
-        android:orientation="vertical">
-
-        <ImageView
-            android:id="@+id/imageView"
-            app:srcCompat="@drawable/no_absence_icon_black"
-            android:layout_width="60dp"
-            android:layout_height="60dp" />
-
-        <TextView
-            android:id="@+id/textView"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="@string/no_absence_on_record"
-            android:textAlignment="center" />
-    </LinearLayout>
-
-    <ListView
-        android:id="@+id/absenceList"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_alignParentStart="true"
-        android:layout_below="@+id/imageButton"
-        android:divider="@null" />
-</RelativeLayout>
