+++ app/src/main/res/values/strings.xml
@@ -74,7 +74,6 @@
     <string name="avggraph_termseparator">End of term</string>
     <string name="select_date">Select Date</string>
     <string name="class_number">#%1$d</string>
-    <string name="no_absence_on_record">You haven\'t missed a single class on this day! (Hooray!)</string>
 
     <!-- Unknown string values -->
     <string name="navigation_drawer_open">Open navigation drawer</string>
